+++ spec/unit/dotnet/parsers/csproj_spec.rb
@@ -1,43 +1,43 @@
 # frozen_string_literal: true
 
 RSpec.describe Spandx::Dotnet::Parsers::Csproj do
-  subject { described_class.new }
+  subject(:described_instance) { described_class.new }
+
+  def build(name, version, path)
+    Spandx::Core::Dependency.new(name: name, version: version, path: path)
+  end
 
   describe '#parse' do
     context 'when parsing a .csproj file' do
-      let(:lockfile) { fixture_file('nuget/example.csproj') }
+      subject { described_instance.parse(file) }
 
-      let(:because) { subject.parse(lockfile) }
-      let(:jive) { because.find { |item| item.name == 'jive' } }
+      let(:file) { fixture_file('nuget/example.csproj') }
 
-      specify { expect(jive.name).to eql('jive') }
-      specify { expect(jive.version).to eql('0.1.0') }
+      let(:jive) { subject.find { |item| item.name == 'jive' } }
+
+      specify { expect(subject).to match_array([build('jive', '0.1.0', file)]) }
     end
 
     context 'when parsing a .csproj file that has a reference to another project' do
-      let(:lockfile) { fixture_file('nuget/nested/test.csproj') }
+      subject { described_instance.parse(fixture_file('nuget/nested/test.csproj')) }
 
-      let(:because) { subject.parse(lockfile) }
-
-      specify { expect(because.map(&:name)).to match_array(%w[jive xunit]) }
+      specify { expect(subject.map(&:name)).to match_array(%w[jive xunit]) }
     end
 
     context 'when parsing `Nancy.Hosting.Self.csproj`' do
-      let(:lockfile) { fixture_file('nuget/Nancy.Hosting.Self.csproj') }
+      subject { described_instance.parse(fixture_file('nuget/Nancy.Hosting.Self.csproj')) }
 
-      let(:because) { subject.parse(lockfile) }
-
-      specify { expect(because.count).to be(1) }
-      specify { expect(because[0].name).to eql('System.Security.Principal.Windows') }
-      specify { expect(because[0].version).to eql('4.3.0') }
+      specify { expect(subject.count).to be(1) }
+      specify { expect(subject[0].name).to eql('System.Security.Principal.Windows') }
+      specify { expect(subject[0].version).to eql('4.3.0') }
     end
   end
 
-  describe '.matches?' do
-    specify { expect(subject.matches?('/root/simple.csproj')).to be(true) }
-    specify { expect(subject.matches?('C:\Documents and Settings\simple.csproj')).to be(true) }
-    specify { expect(subject.matches?('C:\Documents and Settings\hello world.csproj')).to be(true) }
-    specify { expect(subject.matches?('/root/Packages.props')).to be(true) }
-    specify { expect(subject.matches?('/root/simple.sln')).to be(false) }
+  describe '#match?' do
+    specify { is_expected.to be_match(to_path('/root/Packages.props')) }
+    specify { is_expected.to be_match(to_path('/root/simple.csproj')) }
+    specify { is_expected.not_to be_match(to_path('/root/simple.sln')) }
+    specify { is_expected.to be_match(to_path('C:\Documents and Settings\hello world.csproj')) }
+    specify { is_expected.to be_match(to_path('C:\Documents and Settings\simple.csproj')) }
   end
 end