+++ spec/unit/js/parsers/npm_spec.rb
@@ -1,13 +1,19 @@
 # frozen_string_literal: true
 
 RSpec.describe Spandx::Js::Parsers::Npm do
-  subject { described_class.new }
+  describe '#parse' do
+    subject { described_class.new.parse(fixture_file('js/npm/package-lock.json')) }
 
-  describe '#parse lock file' do
-    let(:lockfile) { fixture_file('js/npm/package-lock.json') }
     let(:expected_dependencies) { fixture_file_content('js/npm/expected').lines.map(&:chomp) }
-    let(:result) { subject.parse(lockfile) }
 
-    specify { expect(result.map { |x| "#{x.name}@#{x.version}" }) .to match_array(expected_dependencies) }
+    specify { expect(subject.map { |x| "#{x.name}@#{x.version}" }) .to match_array(expected_dependencies) }
+  end
+
+  describe '#match?' do
+    it { is_expected.to be_match(to_path('package-lock.json')) }
+    it { is_expected.to be_match(to_path('./package-lock.json')) }
+    it { is_expected.to be_match(to_path('/root/package-lock.json')) }
+    it { is_expected.not_to be_match(to_path('sitemap.xml')) }
+    it { is_expected.not_to be_match(to_path('/root/notpackage-lock.json')) }
   end
 end