+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-Version 0.13.3
+Version 0.13.4
 
 # Changelog
 
@@ -8,6 +8,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.13.4] - 2020-05-26
+### Added
+- Add detected file path to report output.
+
+### Changed
+- Use `Pathname` instead of `String` to represent file paths.
+- Scan current directory when a path is not specified.
 
 ## [0.13.3] - 2020-05-19
 ### Fixed